Eliminate logging for multiple addressed rules
Using logging directly to raise this is intrusive, so at most this
should be a warning that is raised and can be handled by checkers.
This proposes to remove it completely, as there are good reasons to
deviate, and guidance like this should be handled at development time,
not runtime.

Signed-off-by: Pierre R. Mai <pmai@pmsf.de>